apply a couple of small optimisation move headers generation to they own pacakges
